+++ app/controllers/documents/search.go
@@ -2,11 +2,11 @@ package controllers
 
 import (
     "github.com/revel/revel"
-    "github.com/asaskevich/govalidator"
-    "github.com/GDG-SSU/wigo/app/models"
-    "github.com/GDG-SSU/wigo/app"
     "log"
+    "github.com/GDG-SSU/wigo/app/models"
     "github.com/GDG-SSU/wigo/app/routes"
+    "github.com/asaskevich/govalidator"
+    "github.com/GDG-SSU/wigo/app"
 )
 
 type SearchDocumentForm struct {
@@ -21,6 +21,7 @@ func parseSearchDocumentForm(p *revel.Params) (SearchDocumentForm, error) {
     return *doc, err
 }
 
+
 func (c Document) Search() revel.Result {
     // When submit button is clicked
     if c.Request.Method == "GET" {
@@ -32,14 +33,16 @@ func (c Document) Search() revel.Result {
 
         var documents []models.Document
 
-        app.DB.Where("Title LIKE ? OR Content LIKE ?", searchDocumentForm.Title, searchDocumentForm.Title).Find(&documents)
+        app.DB.Table("documents").Select("id, title").Where("Title LIKE ? OR Content LIKE ?", "%" + searchDocumentForm.Title + "%", "%" + searchDocumentForm.Title + "%").Find(&documents)
 
         // Check existing document
         if len(documents) == 0 {
             return c.Redirect(routes.Document.Write())
         }
+
         c.RenderArgs["documents"] = documents
         return c.RenderTemplate("Document/search_results.html")
     }
     return c.RenderText("Post is not supported")
 }
+
